Add files via upload
I'm in the middle of changing somethings to the code. Right now I'm not copying from the location of the stored video files, and I also need to make a options button for users can select whether or not the file is for Anime, Movies, or TV. Also need to make the directories if they are none existent.